backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
python-multipart
pdfservices-sdk==4.0.0
pypandoc